Extract RTDB modular ops into host/rtdb.ts
Move rtdb.get/set/update/remove/push/adminSnapshot with the RTDB sentinel
reconstruction and rtdbSnapToWire (exported, still used by the value-sub
handler) into host/rtdb.ts behind isRtdbOp/handleRtdbOp. handleOp routes to it.
Bodies verbatim; orphaned pyric/database/modular op imports dropped from
host.ts. No behavior change.
